backend/pyproject.toml
[tool.poetry]
name = "gccsite-api"
version = "0.1.0"
description = "Backend of Girls Can Code! website"
authors = ["Association Prologin <info@prologin.org>"]
license = "AGPL-3.0"

[[tool.poetry.source]]
name = "gitlab-prologin"
url = "https://gitlab.com/api/v4/groups/prologin/-/packages/pypi/simple"
secondary = true

[tool.poetry.dependencies]
python = "^3.10"
coverage = { version = "*", extras = [ "toml" ] }
django = "^4.1"
django-filter = "^21.1"
gunicorn = "^20.1"
ipython = "*"
jsonschema = "^4.13.0"
prologin-django-utils = { version = "0.1.0", extras = ["auth", "celery-extra", "drf", "redis"] }
prospector = {extras = ["with_everything"], version = "*"}
psycopg2 = "^2.9"


[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
omit = [
    "../*migrations*",
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    # Re-enable the standard pragma
    "pragma: nocover",

    # Don't complain if tests don't hit defensive assertion code:
    "raise NotImplementedError",

    # Don't complain about missing debug-only code:
    "def __repr__",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
]

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
